requirements.txt
astunparse==1.6.3
cachetools==5.4.0
certifi==2022.12.7
cycler==0.11.0
fonttools==4.38.0
ftfy==6.1.1
jclip==1.0
jittor==1.3.8.5
kiwisolver==1.4.5
matplotlib==3.5.3
numpy==1.21.6
nvidia-ml-py==12.535.161
nvitop==1.3.2
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pip==22.3.1
psutil==6.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.4.16
scipy==1.7.3
setuptools==65.6.3
six==1.16.0
termcolor==2.3.0
termplotlib==0.3.9
tqdm==4.66.4
typing_extensions==4.7.1
wcwidth==0.2.13
wheel==0.38.4
